+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.24
 
 - Changed signature of `draw` function to return a reference to the internal `Texture`.
+- Created `TextureStore` for `Widget`s to handle stored `Texture` objects.
 
 ## 0.4.23